(work-in-progress) 	o  changing RTCP XR header format more compliable to the standard (RFC 3611)
again, this commit is still work-in-progress
	1) need to finish rtp/session.cpp +1169
	2) need to have a mean to bit flipping on XR chunk (MSB)
       if MSB = 1, then it is AckVec
	   if MSB = 0, then it is Ack of Ack

+++ cc/tfwc_rcvr.h
@@ -45,15 +45,15 @@ public:
 			u_int16_t ackofack, u_int32_t ts);
 
 protected:
-	inline u_int32_t tfwc_rcvr_getvec() { return tfwcAV; }
+	inline u_int16_t tfwc_rcvr_getvec() { return tfwcAV; }
 	inline u_int32_t tfwc_rcvr_ts_echo() { return ts_echo_; }
-	u_int32_t tfwcAV;	// AckVec (bit vector)
+	u_int16_t tfwcAV;	// AckVec (bit vector)
 	u_int16_t currseq_;	// current sequence number
 	u_int16_t prevseq_;	// previous sequence number
 	u_int16_t ackofack_;	// ackofack
 private:
 	// trim ackvec
-	inline void trimvec(u_int32_t vec, int offset) {
+	inline void trimvec(u_int16_t vec, int offset) {
 		tfwcAV = vec >> offset;
 	}
